backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
pydantic==2.7.1
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
